backend/requirements.txt
Flask==2.1.1
Werkzeug==2.1.2
requests==2.26.0
flask-cors==3.0.10
gunicorn==20.1.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1